Refactor temp schema generator to its own file
